fixed k to int

+++ pylearner/bindings.cpp
@@ -55,7 +55,7 @@ PYBIND11_MODULE(learner_ext, m) {
             int r_val;
             if (r.is_none()) {
                 py::module screenot = py::module::import("screenot");
-                double max_rank = std::min(Y_source.rows(), Y_source.cols()) / 3.0;
+                int max_rank = static_cast<int>(std::min(Y_source.rows(), Y_source.cols()) / 3.0);
                 py::object result_obj = screenot.attr("adaptiveHardThresholding")(Y_source, max_rank);
                 py::tuple result_tuple = result_obj.cast<py::tuple>();
                 r_val = std::max(result_tuple[2].cast<int>(), 1);
@@ -132,7 +132,7 @@ dict
             int r_val;
             if (r.is_none()) {
                 py::module screenot = py::module::import("screenot");
-                double max_rank = std::min(Y_source.rows(), Y_source.cols()) / 3.0;
+                int max_rank = static_cast<int>(std::min(Y_source.rows(), Y_source.cols()) / 3.0);
                 py::object result_obj = screenot.attr("adaptiveHardThresholding")(Y_source, max_rank);
                 py::tuple result_tuple = result_obj.cast<py::tuple>();
                 r_val = std::max(result_tuple[2].cast<int>(), 1);
